pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "greybeard"
version = "0.1.0"
description = "CLI-first Staff Review & Decision Assistant"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Brian Hopkins", email = "btotharye@gmail.com" },
]
keywords = [
    "code-review",
    "llm",
    "staff-engineer",
    "architecture",
    "mcp",
    "cli",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/btotharye/greybeard"
Documentation = "https://greybeard.readthedocs.io/en/latest/"
"Source Code" = "https://github.com/btotharye/greybeard"
"Issue Tracker" = "https://github.com/btotharye/greybeard/issues"

[project.scripts]
greybeard = "staff_review.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
anthropic = [
    "anthropic>=0.25",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pymdown-extensions>=10.0",
]
all = [
    "anthropic>=0.25",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pymdown-extensions>=10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["staff_review"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
